+++ src/jobs/triggered/reportFlowTelemetry/run.ts
@@ -1,0 +1,17 @@
+/**
+ * 測定データ報告
+ */
+import * as createDebug from 'debug';
+import * as Controller from '../../../controller/reportFlowTelemetry';
+
+const debug = createDebug('sskts-monitoring-jobs');
+
+Controller.main()
+    .then(() => {
+        debug('success!');
+    })
+    .catch((err) => {
+        // tslint:disable-next-line:no-console
+        console.error(err);
+        process.exit(1);
+    });