+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
-- install @motionpicture/sskts-domain@29.1.5
+- install @motionpicture/sskts-domain@30.0.0
 
 ### Deprecated